+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 5.5.1:
+ - Issue claims with byte arrays without encoding them to Base64
  - Update to VC-K 5.5.1
  - define a revocation identifier for non-empty MDOC credentials
  - expose revocation identifier